+++ food_app/index.css
@@ -29,15 +29,22 @@
     padding: 5px;
     width: 185px;
     margin: 9px ;
+    transform: scale(1);
+  transition: transform 0.3s ease-in-out;
 }
 
 .res-card:hover {
-    border: 1px solid black;
+    border-radius:10px;
+    transform: scale(.95);
+    transition: box-shadow .5s;
+    box-shadow: 0px 0px 10px 5px #ccc;
 }
 
-/* .res-card > h4, h5 {
-    margin: auto;
-} */
+.res-card-details{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+}
 
 .search{
     padding: 10px;
@@ -47,5 +54,24 @@
     width: 100%;
     height: 130px;
     object-fit: cover;
+    border-radius: 10px;
+}
+
+.filter-btn{
+    margin: 10px 10px;
+    border-radius: 18px;
+    background-color: rgb(255, 255, 255);
+    border: 1px solid rgb(226, 226, 231);
+    padding: 9px 12px;
+    width: auto;
+    cursor: pointer;
+    box-shadow: rgba(2, 6, 12, 0.04) 0px 2px 12px;
+    transform: scale(1);
+    transition: background-color 0.1s ease-in 0s, border-color 0.1s ease-in 0s, box-shadow 0.1s ease-in 0s, width 0.2s cubic-bezier(0.65, 0, 0.35, 1) 0s;
+}
+
+.filter-btn:hover{
+    transform: scale(.95);
+    transition: box-shadow .5s;
 }
