+++ food_app/index.html
@@ -18,6 +18,6 @@
       </script>
     </div>
 
-    <script type="module" src="./App.js"></script>
+    <script type="module" src="./src/App.js"></script>
   </body>
 </html>
